13/09/2021

+++ src/main/java/com/anodiam/CRUDStudentProfile/db/repository/PermissionRepository.java
@@ -4,6 +4,8 @@ import com.anodiam.CRUDStudentProfile.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, BigInteger> {
 }
