+++ src/main/java/com/anodiam/CRUDStudentProfile/db/repository/RoleRepository.java
@@ -4,6 +4,8 @@ import com.anodiam.CRUDStudentProfile.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, BigInteger> {
 }
